NBodyProblem.BurnesAndHut

Description

The simple Haskell implementation of a Barnes-Hut galaxy simulator.

Documentation

data Node

A node represents a body if it is an endnote (i.e. if node.child is None) or an abstract node of the quad-tree if it has child. Instead of storing the position of a node, we store the mass times position, m_pos. This makes it easier to update the center-of-mass.

Constructors

Node

Fields

m :: Float

m_pos :: [Float]

momentum :: [Float]

s :: Float

relpos :: [Float]

child :: Maybe [Maybe Node]

initNode :: Float -> Float -> Float -> Node

The initializer creates a child-less node (an actual body).

intoNextQuadrant :: Node -> ([Node], Int)

Places node into next-level quadrant and returns the quadrant number.

_subdivide :: Node -> Int -> Int -> [Node]

Places node into next-level quadrant along direction i and recomputes the relative position relpos of the node inside this quadrant.

pos :: Node -> [Float]

Physical position of node, independent of currently active quadrant.

resetTo0thQuadrant :: Node -> Node

Re-positions the node to the level-0 quadrant (full domain).

dist :: Node -> Node -> Float

Distance between present node and another node.

forceOn :: Node -> Node -> Float -> [Float]

Force which the present node is exerting on a given body.

add :: Node -> Maybe Node -> Node

Barnes-Hut algorithm: Creation of the quad-tree. This function adds a new body into a quad-tree node. Returns an updated version of the node.

fromJust :: Maybe a -> a

isNothing :: Maybe a -> Bool

verlet :: [Node] -> Node -> Float -> Float -> Float -> [Node]

Execute a time iteration according to the Verlet algorithm.